Completed assignment 11 and Fundamentals
All Quizzes are also completed. If I missed one, Sorry, and its probably
because it is hard to keep track of which ones have been completed when
when browser history is erased, or different computers are used. I think
I got them all. The memory game is working to the requirements as best
as I can tell.